refactor: Simplify CI workflow steps for formatting and linting
This commit simplifies the CI workflow steps for formatting and linting by replacing the 'actions-rs/cargo' action with direct 'cargo' commands. The 'cargo fmt' step now uses the 'run' command with the 'cargo fmt --all -- --check' arguments, and the 'cargo clippy' step uses the 'run' command with the 'cargo clippy --verbose' arguments. This change reduces the complexity of the workflow and improves readability.

Signed-off-by: luong-hq <luong@hyperquery.ai>
GitOrigin-RevId: af6452a542317f04ed652ae84219c4c37c3e4fed